+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# dev
+- create_random_laz: add classifications parameter
+
 # 1.12.3
 - [fix] add_points_in_pointcloud: format consistency when input shapefile does not have a crs