+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # dev
 - create_random_laz: add classifications parameter
+- las_rename_dimension: if dimension doesn't exist, do not raise an error but log a warning
 
 # 1.12.3
 - [fix] add_points_in_pointcloud: format consistency when input shapefile does not have a crs